use existsSync

+++ test/functional/fixtures/run-options/stop-on-first-fail/test.js
@@ -13,17 +13,12 @@ const getTestRunCount = () => {
 
 describe('Stop test task on first failed test', () => {
     afterEach(() => {
-        try {
-            // NOTE: after mocha is updated to `^7.1.1` the `afterEach` hook is called if the test is skipped/pending
-            // before the update the `afterEach` hook was not called.
-            // When we run the test not in `chrome` the file will not exist, since we have the { only: 'chrome' } option,
-            // so we cannot unlink it.
+        // NOTE: after mocha is updated to `^7.1.1` the `afterEach` hook is called if the test is skipped/pending
+        // before the update the `afterEach` hook was not called.
+        // When we run the test not in `chrome` the file will not exist, since we have the { only: 'chrome' } option,
+        // so we cannot unlink it.
+        if (fs.existsSync(TEST_RUN_COUNT_FILENAME))
             fs.unlinkSync(TEST_RUN_COUNT_FILENAME);
-        }
-        /*eslint-disable no-empty */
-        catch (err) {
-        }
-        /*eslint-disable no-empty */
     });
 
     it('Basic', () => {
